requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0 